Crate dkg_primitives

Re-exports

pub use serde_json;

Modules

crypto

handlers

keys

offchain

proposal

rounds

traits

types

util

Utility functions for Ethereum.

utils

utils

Structs

AccessListItem

Account

AggregatedMisbehaviourReports

AggregatedPublicKeys

AuthoritySet

Block

Bloom

Bloom hash type with 256 bytes (2048 bits) size.

BloomRef

Commitment

EIP1559Transaction

EIP1559TransactionMessage

EIP2930Transaction

EIP2930TransactionMessage

FromStrRadixErr

The error type for parsing numbers from strings.

FunctionSignature

Proposal Target Function Signature (4 bytes).

H32

H64

H128

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H264

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

H520

Header

Ethereum header definition.

LegacyTransaction

LegacyTransactionMessage

Log

OffchainSignedProposals

PartialHeader

Partial header definition without ommers hash and transactions root.

ProposalHeader

Proposal Header (40 bytes).

ProposalNonce

Proposal Nonce (4 bytes).

Receipt

RefreshProposal

RefreshProposalSigned

ResourceId

Proposal Target ResourceId (32 bytes).

TransactionRecoveryId

TransactionSignature

U64

Little-endian large integer type

U128

Little-endian large integer type

U256

Little-endian large integer type

U512

Little-endian large integer type

UnsignedProposal

Enums

BloomInput

ConsensusLog

DKGPayloadKey

DKGReport

FromDecStrErr

Conversion from decimal string error

FromStrRadixErrKind

A list of error categories encountered when parsing numbers.

Proposal

ProposalAction

ProposalKind

TransactionAction

TransactionV1

TransactionV2

TypedChainId

Proposal Target Chain and its type (6 bytes).

Constants

DKG_ENGINE_ID

GENESIS_AUTHORITY_SET_ID

Authority set id starts with zero at genesis

GENESIS_BLOCK_NUMBER

KEY_TYPE

PROPOSAL_SIGNATURE_LENGTH

UNTRACK_INTERVAL

Traits

BigEndianHash

DKGApi

ProposalHandlerTrait

Functions

keccak_256

Utility fn to calculate keccak 256 has

Type Definitions

AccessList

Address

AuthorityIndex

AuthoritySetId

BlockV0

BlockV1

BlockV2

MmrRootHash

The type used to represent an MMR root hash.

Public

PublicKeyAndSignature

Secret

Signature

TransactionV0