Module dkg_primitives::rounds::sign::gg20_sign

High-level threshold signing protocol implementation

Key feature of GG20 protocol is one-round online signing, meaning that every party needs to broadcast just a single message to sign a data. However, it still requires completing an offline computation for fixed set of parties

How to get things work

First of all, parties need to carry out distributed key generation protocol (see keygen module). After DKG is successfully completed, it outputs LocalKey — a party local secret share. Then you fix a set of parties who will participate in threshold signing, and they run OfflineStage protocol. OfflineStage implements StateMachine and can be executed in the same way as Keygen. OfflineStage outputs a CompletedOfflineStage. SignManual takes a CompletedOfflineStage and allows you to perform one-round signing. It doesn’t implement StateMachine, but rather provides methods to construct messages and final signature manually (refer to SignManual documentation to see how to use it).

Modules

rounds

Structs

CompletedOfflineStage

OfflineProtocolMessage

OfflineStage

Offline Stage of GG20 signing

PartialSignature

SignManual

Manual GG20 signing

Enums

Error

InternalError

ProceedError

SignError