Trait dkg_primitives::DKGApi

pub trait DKGApi<Block, AuthorityId, N>: Core<Block> where
    Block: Block,
    AuthorityId: Codec + PartialEq<AuthorityId>,
    N: Codec + PartialEq<N> + AtLeast32BitUnsigned, {

    fn authority_set(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }

    fn authority_set_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }

    fn signature_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }

    fn signature_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }

    fn keygen_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }

    fn keygen_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }

    fn next_signature_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }

    fn next_signature_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }

    fn next_keygen_threshold(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u16, ApiError> { ... }

    fn next_keygen_threshold_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u16, ApiError> { ... }

    fn queued_authority_set(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }

    fn queued_authority_set_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<AuthoritySet<AuthorityId>, ApiError> { ... }

    fn should_refresh(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        _block_number: N
    ) -> Result<bool, ApiError> { ... }

    fn should_refresh_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        _block_number: N
    ) -> Result<bool, ApiError> { ... }

    fn next_dkg_pub_key(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<Vec<u8, Global>>, ApiError> { ... }

    fn next_dkg_pub_key_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<Vec<u8, Global>>, ApiError> { ... }

    fn dkg_pub_key(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<Vec<u8, Global>>, ApiError> { ... }

    fn dkg_pub_key_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<Vec<u8, Global>>, ApiError> { ... }

    fn get_unsigned_proposals(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Vec<UnsignedProposal, Global>, ApiError> { ... }

    fn get_unsigned_proposals_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Vec<UnsignedProposal, Global>, ApiError> { ... }

    fn get_max_extrinsic_delay(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        block_number: N
    ) -> Result<N, ApiError> { ... }

    fn get_max_extrinsic_delay_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        block_number: N
    ) -> Result<N, ApiError> { ... }

    fn get_authority_accounts(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>), ApiError> { ... }

    fn get_authority_accounts_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>), ApiError> { ... }

    fn get_reputations(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        authorities: Vec<AuthorityId, Global>
    ) -> Result<Vec<(AuthorityId, i64), Global>, ApiError> { ... }

    fn get_reputations_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext, 
        authorities: Vec<AuthorityId, Global>
    ) -> Result<Vec<(AuthorityId, i64), Global>, ApiError> { ... }

    fn next_pub_key_sig(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<Option<Vec<u8, Global>>, ApiError> { ... }

    fn next_pub_key_sig_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<Option<Vec<u8, Global>>, ApiError> { ... }

    fn refresh_nonce(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<u32, ApiError> { ... }

    fn refresh_nonce_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<u32, ApiError> { ... }

    fn time_to_restart(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>
    ) -> Result<N, ApiError> { ... }

    fn time_to_restart_with_context(
        &self, 
        __runtime_api_at_param__: &BlockId<Block>, 
        context: ExecutionContext
    ) -> Result<N, ApiError> { ... }

}

Provided methods

fn authority_set(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the current active authority set

fn authority_set_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the current active authority set

fn signature_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the current signature threshold for the DKG

fn signature_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the current signature threshold for the DKG

fn keygen_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the current keygen threshold for the DKG

fn keygen_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the current keygen threshold for the DKG

fn next_signature_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the next signature threshold for the DKG

fn next_signature_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the next signature threshold for the DKG

fn next_keygen_threshold(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u16, ApiError>

Return the next keygen threshold for the DKG

fn next_keygen_threshold_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u16, ApiError>

Return the next keygen threshold for the DKG

fn queued_authority_set(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the next authorities active authority set

fn queued_authority_set_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<AuthoritySet<AuthorityId>, ApiError>

Return the next authorities active authority set

fn should_refresh(
&self,
__runtime_api_at_param__: &BlockId<Block>,
_block_number: N
) -> Result<bool, ApiError>

Check if refresh process should start

fn should_refresh_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
_block_number: N
) -> Result<bool, ApiError>

Check if refresh process should start

fn next_dkg_pub_key(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<Vec<u8, Global>>, ApiError>

Fetch DKG public key for queued authorities

fn next_dkg_pub_key_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<Vec<u8, Global>>, ApiError>

Fetch DKG public key for queued authorities

fn dkg_pub_key(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<Vec<u8, Global>>, ApiError>

Fetch DKG public key for current authorities

fn dkg_pub_key_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<Vec<u8, Global>>, ApiError>

Fetch DKG public key for current authorities

fn get_unsigned_proposals(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Vec<UnsignedProposal, Global>, ApiError>

Get list of unsigned proposals

fn get_unsigned_proposals_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Vec<UnsignedProposal, Global>, ApiError>

Get list of unsigned proposals

fn get_max_extrinsic_delay(
&self,
__runtime_api_at_param__: &BlockId<Block>,
block_number: N
) -> Result<N, ApiError>

Get maximum delay before which an offchain extrinsic should be submitted

fn get_max_extrinsic_delay_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
block_number: N
) -> Result<N, ApiError>

Get maximum delay before which an offchain extrinsic should be submitted

fn get_authority_accounts(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>), ApiError>

Current and Queued Authority Account Ids [/current_authorities/, /next_authorities/]

fn get_authority_accounts_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<(Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>, Vec<<<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId, Global>), ApiError>

Current and Queued Authority Account Ids [/current_authorities/, /next_authorities/]

fn get_reputations(
&self,
__runtime_api_at_param__: &BlockId<Block>,
authorities: Vec<AuthorityId, Global>
) -> Result<Vec<(AuthorityId, i64), Global>, ApiError>

Reputations for authorities

fn get_reputations_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext,
authorities: Vec<AuthorityId, Global>
) -> Result<Vec<(AuthorityId, i64), Global>, ApiError>

Reputations for authorities

fn next_pub_key_sig(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<Option<Vec<u8, Global>>, ApiError>

Fetch DKG public key for sig

fn next_pub_key_sig_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<Option<Vec<u8, Global>>, ApiError>

Fetch DKG public key for sig

fn refresh_nonce(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<u32, ApiError>

Get next nonce value for refresh proposal

fn refresh_nonce_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<u32, ApiError>

Get next nonce value for refresh proposal

fn time_to_restart(
&self,
__runtime_api_at_param__: &BlockId<Block>
) -> Result<N, ApiError>

Get the time to restart for the dkg keygen

fn time_to_restart_with_context(
&self,
__runtime_api_at_param__: &BlockId<Block>,
context: ExecutionContext
) -> Result<N, ApiError>

Get the time to restart for the dkg keygen