Trait dkg_primitives::traits::OnDKGPublicKeyChangeHandler

pub trait OnDKGPublicKeyChangeHandler<AuthoritySetId> where
    AuthoritySetId: Copy, {
    fn on_dkg_public_key_changed(
        authority_id: AuthoritySetId, 
        dkg_public_key: Vec<u8, Global>
    ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>;
}

A trait for when the DKG Public Key get changed.

This is used to notify the runtime that the DKG signer has changed. for example, this could be used to know that we should trigger a re-signing of any pending signed proposals to take into account the new DKG signer.

Required methods

fn on_dkg_public_key_changed(
authority_id: AuthoritySetId,
dkg_public_key: Vec<u8, Global>
) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>