Module dkg_primitives::utils

Structs

StoredLocalKey

Constants

DKG_LOCAL_KEY_FILE

QUEUED_DKG_LOCAL_KEY_FILE

Functions

cleanup

convert_signature

decrypt_data

Decrypt an encrypted data with a secret_key_bytes using XChaCha20Poly1305 cipher.

encrypt_data

Encrypt a vector of bytes data with a secret_key.

get_from_seed

Helper function to generate a crypto pair from seed

insert_controller_account_keys_into_keystore

Inserts a key of type ACCOUNT into the keystore for development/testing.

select_random_set

Select a random subset of unsigned u16 from a vector of u16s of size amount with a random seed of seed.

vec_usize_to_u16