Crate dkg_primitives

Re-exports

pub use serde_json;

Modules

crypto

dkg_key_cli

ethereum_abi

gossip_messages

handlers

keys

offchain

proposal

traits

types

utils

Macros

if_ethbloom

Conditional compilation depending on whether ethereum-types is built with ethbloom support.

Structs

AccessListItem

Account

AggregatedMisbehaviourReports

AggregatedProposerVotes

AggregatedPublicKeys

AuthoritySet

Block

Bloom

Bloom hash type with 256 bytes (2048 bits) size.

BloomRef

Commitment

CustomU32Getter

A custom get Struct to allow to import runtime values into the dkg gadget

DKGSignedPayload

EIP658ReceiptData

EIP1559Transaction

EIP1559TransactionMessage

EIP2930Transaction

EIP2930TransactionMessage

FromStrRadixErr

The error type for parsing numbers from strings.

FrontierReceiptData

FunctionSignature

Proposal Target Function Signature (4 bytes).

H32

H64

H128

Fixed-size uninterpreted hash type with 16 bytes (128 bits) size.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H264

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

H520

Header

Ethereum header definition.

LegacyTransaction

LegacyTransactionMessage

Log

OffchainSignedProposalBatches

PartialHeader

Partial header definition without ommers hash and transactions root.

ProposalHeader

Proposal Header (40 bytes).

ProposalNonce

Proposal Nonce (4 bytes).

RefreshProposal

ResourceId

Proposal Target ResourceId (32 bytes).

SignedProposalBatch

An unsigned proposal represented in pallet storage We store the creation timestamp to purge expired proposals

StoredUnsignedProposalBatch

An unsigned proposal represented in pallet storage We store the creation timestamp to purge expired proposals

TransactionRecoveryId

TransactionSignature

U64

Little-endian large integer type Unsigned 64-bit integer.

U128

Little-endian large integer type 128-bit unsigned integer.

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

UnsignedProposal

Enums

BloomInput

ConsensusLog

DKGPayloadKey

DKGReport

EnvelopedDecoderError

DecoderError for typed transactions.

FromDecStrErr

Conversion from decimal string error

FromStrRadixErrKind

A list of error categories encountered when parsing numbers.

MisbehaviourType

Proposal

Proposal enum

ProposalAction

ProposalKind

Proposal kind enum

ReceiptAny

ReceiptV2

ReceiptV3

TransactionAction

TransactionV1

TransactionV2

TypedChainId

Proposal Target Chain and its type (6 bytes).

Constants

ASSOCIATED_BLOCK_ID_MESSAGE_DELIVERY_TOLERANCE

So long as the associated block id is within this tolerance, we consider the message as deliverable. This should be less than the SIGN_TIMEOUT

DKG_ENGINE_ID

Engine ID for DKG

GENESIS_AUTHORITY_SET_ID

Authority set id starts with zero at genesis

KEYGEN_TIMEOUT

The keygen timeout limit in blocks before we consider misbehaviours

KEY_TYPE

Key type for DKG keys

PROPOSAL_SIGNATURE_LENGTH

SIGN_TIMEOUT

The sign timeout limit in blocks before we consider proposal as stalled

UNTRACK_INTERVAL

Traits

BigEndianHash

DKGApi

EnvelopedDecodable

Decodable typed transactions.

EnvelopedEncodable

Encodable typed transactions.

ProposalHandlerTrait

Functions

associated_block_id_acceptable

keccak_256

Utility fn to calculate keccak 256

Type Definitions

AccessList

Address

AuthorityIndex

AuthoritySetId

BatchId

BlockAny

BlockV0

BlockV1

BlockV2

EIP1559ReceiptData

EIP2930ReceiptData

MaxAuthorities

Max authorities

MaxKeyLength

Max size for keys

MaxProposalLength

MaxProposalsInBatch

Max proposals in a batch

MaxReporters

Max reporters

MaxResources

Max resources to store onchain

MaxSignatureLength

Max size for signatures

MaxVoteLength

Max size for vote proposal A vote consists of a new address, a leaf index, and a merkle proof [20 bytes, 4 bytes, 32 * HEIGHT bytes] If we want to support up to 1000 validators than the max vote length is 20 + 4 + 32 * 10 = 364 bytes or in bits 2912

MaxVotes

Max votes to store onchain

MmrRootHash

The type used to represent an MMR root hash.

ProposerVote

Proposer vote type

Public

PublicKeyAndSignature

ReceiptV0

ReceiptV1

Secret

SessionId

A typedef for keygen set / session id

Signature

SignerSetId

Signer set ID

TransactionAny

TransactionV0